Moved EntityModel and ClassModel classes to knowledge_graph_model.py, fixed annotator.py, column_classifier.py, parser.py, tabular_data_model.py and main.py
